Add node

+++ 0x12-singly_linked_lists/2-add_node.c
@@ -1,7 +1,5 @@
 #include "lists.h"
 #include <stdlib.h>
-#include <string.h>
-#include <stdio.h>
 
 /**
  * add_node - adds a new node at the begining of a list
@@ -13,7 +11,7 @@
 list_t *add_node(list_t **head, const char *str)
 {
 	list_t *new;
-	unsigned long int i;
+	size_t i;
 
 	new = malloc(sizeof(list_t));
 	if (new == NULL)
